requirements.txt
# Web Frameworks
fastapi==0.115.9
uvicorn==0.34.3
starlette==0.45.3
gradio==5.32.1
streamlit==1.45.1

# Auth, Security, Env
bcrypt==4.3.0
python-dotenv==1.1.0
python-multipart==0.0.20
cryptography==45.0.3

# ORM and DB
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
mysql-connector-python==9.3.0
pymongo==4.13.0

# Celery, Redis
celery==5.5.3
redis==6.2.0
billiard==4.2.1
vine==5.1.0

# LLMs + Embedding Providers
openai==1.84.0
google-genai==1.18.0
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.0
tenacity==9.1.2

# Vector Stores
faiss-cpu==1.11.0
chromadb==1.0.12
milvus-lite==2.4.12
qdrant-client==1.14.2
pinecone-client==6.0.0

# Data Processing
pandas==2.2.3
numpy==2.2.6
scikit-learn==1.6.1
scipy==1.15.3
PyYAML==6.0.2
pathlib==1.0.1
joblib==1.5.1

# File I/O & Parsing
pdfminer.six==20250327
pdfplumber==0.11.6
PyMuPDF==1.26.0
openpyxl==3.1.5
ujson==5.10.0
pydub==0.25.1
pyarrow==20.0.0

# API Utils
httpx==0.28.1
requests==2.32.3
requests-oauthlib==2.0.0
orjson==3.10.18

# Serialization & Pydantic
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
dataclasses-json==0.6.7

# Background, Workers, Monitoring
watchdog==6.0.0
watchfiles==1.0.5

# Markdown & Templates
Jinja2==3.1.6
markdown-it-py==3.0.0
mdurl==0.1.2

# Miscellaneous
huggingface-hub==0.32.4
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
langsmith==0.3.44
coloredlogs==15.0.1
humanfriendly==10.0
protobuf==5.29.5
click==8.2.1
rich==14.0.0
regex==2024.11.6
typing_extensions==4.14.0
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]